Fixed entity inheritance. Adjusted tests
